it works now, still using test db
I really really want to find someway to get the server used by speedtest
to include, but the --simple flag doesn't give that info. without the
--simple flag the output is not simple to parse

if you specify the server and it's down, you are out of luck. If you
don't specify the server it will perhaps use a different server,
without you knowing.